change type casting

+++ pyopencl_tutorial/kernel.cl
@@ -1,7 +1,7 @@
 #pragma OPENCL EXTENSION cl_khr_fp16 : enable
 #pragma OPENCL EXTENSION cl_arm_printf : enable
 
-__kernel void add1(__global float* a, __global float* b, __global float* c, __global float* d, __global float* e) 
+__kernel void add1(__global float* a, __global float* b, __global float* c, __global half* d, __global half* e) 
 { 
 	__global half* a1 = (__global half*)a;
 	__global half* b1 = (__global half*)b;
